% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.ndp.R
\name{as.array.stapDP}
\alias{as.array.stapDP}
\title{Retrieve NDP parameter samples in Array Form}
\usage{
\method{as.array}{stapDP}(x, ...)
}
\arguments{
\item{x}{a ndp object}

\item{...}{ignored}
}
\description{
Retrieve NDP parameter samples in Array Form
}
\seealso{
as.matrix.ndp for a matrix container of *all* parameters
}
