% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.ndp.R
\name{as.matrix.ndp}
\alias{as.matrix.ndp}
\title{Retrieve Parameter Samples in Matrix Form}
\usage{
\method{as.matrix}{ndp}(x, ...)
}
\arguments{
\item{x}{a ndp object}

\item{...}{ignored}
}
\description{
Returns the parameter samples from the model associated with the 
means, variances, concentration parameters and NDP probabilities
as  a S X P matrix where S is the number of samples and P is the total number of parameters
}
\seealso{
as.array.stapDP for an array like container for the NDP components
}
