% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cluster_mode.R
\name{assign_mode}
\alias{assign_mode}
\title{Estimates Posterior Mode Cluster Assignment}
\usage{
\method{assign_mode}{ndp}(x, ics = NULL)
}
\arguments{
\item{x}{ndp object}
}
\value{
vector of cluster assignments
}
\description{
Estimates Posterior Mode Cluster Assignment
}
