% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bend.R
\name{bend}
\alias{bend}
\title{Estimate Clustered Inhomgogenous poisson process intensity functions from grouped data using a normal kernel}
\usage{
bend(
  formula,
  benvo,
  base_measure = "Normal",
  mu_0 = 0,
  kappa_0 = 1,
  nu_0 = 1,
  sigma_0 = 1,
  L = 4,
  K = 4,
  a_alpha = 1,
  b_alpha = 1,
  a_rho = 1,
  b_rho = 1,
  iter_max = 5000,
  burn_in = floor(iter_max/2),
  thin = 1,
  seed = NULL
)
}
\arguments{
\item{formula}{bendr formula see the details}

\item{benvo}{\code{\link[rBenvo]{Benvo}} object}

\item{base_measure}{one of c("Normal","Beta") indicating which base measure to use.}

\item{mu_0}{mean hyperparameter for mu normal base measure. Default is 0; Normal(0,1).}

\item{kappa_0}{variance hyperparameter for mu normal base measure. Default is 1; Normal(0,1).}

\item{nu_0}{df hyperparameter for sigma inv chisq base measure. Default is 1; InvChisq(1,1);}

\item{sigma_0}{scale hyperparameter for sigma inv chisq base measure. Default is 1; InvChisq(1,1);}

\item{L}{component truncation number}

\item{K}{intensity cluster truncation number}

\item{a_alpha}{shape hyperparameter for alpha gamma prior}

\item{b_alpha}{scale hyperparameter for alpha gamma prior}

\item{a_rho}{shape hyperparameter for rho gamma prior}

\item{b_rho}{scale hyperparameter for rho gamma prior}

\item{iter_max}{total number of iterations for which to run sampler}

\item{burn_in}{number of iterations to discard as sampler approaches the stationary distribution}

\item{thin}{number of iterations to thin by}

\item{seed}{integer with which to initialize random number generator}
}
\description{
Estimate Clustered Inhomgogenous poisson process intensity functions from grouped data using a normal kernel
}
