% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_measure.R
\name{beta_measure}
\alias{beta_measure}
\title{Beta base measure}
\usage{
beta_measure(a_0 = 1, b_0 = 1, mu_sd = 1, tau_sd = 1)
}
\arguments{
\item{a_0}{hyperparameter for mu beta base measure. Default is 1; uniform(0,1).}

\item{b_0}{hyperparameter for mu beta base measure. Default is 1; uniform(0,1).}

\item{mu_sd}{proposal scale for mus}

\item{tau_sd}{proposal scale for mus}
}
\description{
Beta base measure
}
