% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_nd_nhpp.R
\name{beta_nd_nhpp}
\alias{beta_nd_nhpp}
\title{Estimate the nonhomgogenous poisson process intensity function from grouped data using a beta base measure}
\usage{
beta_nd_nhpp(r, n_j, mu_sd, tau_sd, L = 4, K = 4, a_0 = 1, b_0 = 1,
  a_alpha = 1, b_alpha = 1, a_rho = 1, b_rho = 1, iter_max,
  warm_up, thin = 1, multiple_taus = FALSE, seed = NULL)
}
\arguments{
\item{r}{vector of distances  associated with differing groups}

\item{n_j}{matrix of integers denoting the start and length of each school's associated BEF distances}

\item{mu_sd}{proposal scale for mus}

\item{tau_sd}{proposal scale for mus}

\item{L}{component truncation number}

\item{K}{intensity cluster truncation number}

\item{a_0}{hyperparameter for mu beta base measure. Default is 1; uniform(0,1).}

\item{b_0}{hyperparameter for mu beta base measure. Default is 1; uniform(0,1).}

\item{a_alpha}{hyperparameter for alpha gamma prior}

\item{b_alpha}{hyperparameter for alpha gamma prior}

\item{a_rho}{hyperparameter for rho gamma prior}

\item{b_rho}{hyperparameter for rho gamma prior}

\item{iter_max}{total number of iterations for which to run sampler}

\item{warm_up}{number of iterations for which to burn-in or "warm-up" sampler}

\item{thin}{number of iterations to thin by}

\item{multiple_taus}{logical indicator for whether or not cluster specific scales (tau) should be estimated}

\item{seed}{integer with which to initialize random number generator}
}
\description{
Estimate the nonhomgogenous poisson process intensity function from grouped data using a beta base measure
}
