% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{beta_nd_nhpp_fit}
\alias{beta_nd_nhpp_fit}
\title{Estimate the nonhomgogenous poisson process intensity function from grouped data}
\usage{
beta_nd_nhpp_fit(
  r,
  n_j,
  d,
  mu_sd,
  tau_sd,
  L,
  K,
  J,
  a_0,
  b_0,
  a_alpha,
  b_alpha,
  a_rho,
  b_rho,
  iter_max,
  warm_up,
  thin,
  seed,
  chain,
  num_posterior_samples
)
}
\arguments{
\item{r}{vector of distances associatd with different BEFs}

\item{n_j}{matrix of integers denoting the start and length of each school's associated BEF distances}

\item{d}{a 1D grid of positive real values over which the differing intensities are evaluated}

\item{mu_sd}{scale for mu proposal dist'n}

\item{tau_sd}{scale for tau proposal dist'n}

\item{L}{component truncation number}

\item{K}{intensity cluster truncation number}

\item{J}{number of rows in r matrix; number of groups}

\item{a_0}{hyperparameter for mu base measure}

\item{b_0}{hyperparameter for mu base measure}

\item{a_alpha}{hyperparameter for alpha gamma prior}

\item{b_alpha}{hyperparameter for alpha gamma prior}

\item{a_rho}{hyperparameter for rho gamma prior}

\item{b_rho}{hyperparameter for rho gamma prior}

\item{iter_max}{total number of iterations for which to run sampler}

\item{warm_up}{number of iterations for which to burn-in or "warm-up" sampler}

\item{thin}{number of iterations to thin by}

\item{seed}{integer with which to initialize random number generator}

\item{chain}{integer chain label}

\item{num_posterior_samples}{the total number of posterior samples after burn in}
}
\description{
Estimate the nonhomgogenous poisson process intensity function from grouped data
}
