% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bndp.R
\name{bndp}
\alias{bndp}
\title{Create a bndp object}
\usage{
bndp(object, chains, J)
}
\arguments{
\item{object}{A list provided by the nd_nhpp function}

\item{chains}{number of chains}

\item{J}{number of groups}
}
\value{
An ndp object
}
\description{
Create a bndp object
}
