% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{calc_ci}
\alias{calc_ci}
\title{Calculate the confidence intervals}
\usage{
calc_ci(m.formula, x.values, curvefit, curve.predict, dof, level = 0.95,
  debug = FALSE)
}
\arguments{
\item{m.formula}{model formula}

\item{x.values}{concentration values at which the curve will be plotted}

\item{curvefit}{model fit object from nonlinear least squares}

\item{curve.predict}{predicted effect values at the concentrations specified in x.values}

\item{dof}{degrees of freedom for the inverse student's t distribution ; default is (# measurements)-2}

\item{level}{Level of confidence interval to use (0.95 by default). [0,1]}

\item{debug}{Print and plot intermediate results for debugging}
}
\value{
Confidence interval values for all concentrations in x.values
}
\description{
confidence intervals are calculated beyond the concentrations, at which measurements were taken
}
