% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calc_xrange}
\alias{calc_xrange}
\title{Calculate a vector of x-values for plotting}
\usage{
calc_xrange(m.formula, fdata)
}
\arguments{
\item{m.formula}{model formula}

\item{fdata}{dataframe with concentration values}
}
\description{
Calculate a vector of x-values for plotting
}
