% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cleanAndPivot}
\alias{cleanAndPivot}
\title{Remove NA columns from dataframe and convert to long format}
\usage{
cleanAndPivot(df)
}
\arguments{
\item{df}{dataframe with dose response data. The first column needs to be the concentration.}
}
\value{
long dataframe
}
\description{
Remove NA columns from dataframe and convert to long format
}
