% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndp-methods.R
\name{clustercb}
\alias{clustercb}
\title{Cluster Credible Ball}
\usage{
clustercb(object)
}
\arguments{
\item{object}{an ndp object}
}
\description{
Calculates the credible ball (analog to the credible interval)  for the cluster assignments
}
\seealso{
Wade, Sara, and Zoubin Ghahramani. "Bayesian cluster analysis: Point estimation and credible balls (with discussion)." Bayesian Analysis 13.2 (2018): 559-626.
}
