% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{drfit}
\alias{drfit}
\title{Dose Response curve fit}
\usage{
drfit(m.formula, fdata, level = 0.95, start = vector(),
  verbose = FALSE)
}
\arguments{
\item{m.formula}{Model formula.}

\item{fdata}{Data to fit the model to.}

\item{level}{Level of confidence interval to use (0.95 by default). [0,1]}

\item{start}{initial parameter values for the dose response fitting. c(logEC50,slope)}

\item{verbose}{logical, prints intermediate results if true}
}
\description{
Dose Response curve fit
}
