% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNTC.R
\name{findNTC}
\alias{findNTC}
\title{Find the non-toxic concentration.}
\usage{
findNTC(bendrObj)
}
\arguments{
\item{bendrObj}{A bendr result object}
}
\value{
The absolute (non-logged) non-toxic concentration value
}
\description{
If you use the result of this calculation in your research, please cite Stadnicka-Michalak et al.
DOI: 10.14573/altex.1701231
}
