% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cluster_mode.R
\name{get_error}
\alias{get_error}
\title{Calculates Error Distribution}
\usage{
\method{get_error}{ndp}(x, ics = NULL)
}
\arguments{
\item{x}{ndp object}
}
\value{
vector of error corresponding to specific iteration's cluster configuration
}
\description{
Calculates Error Distribution
}
