% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_mode.R
\name{get_square_error}
\alias{get_square_error}
\title{Calculates Error Distribution under square loss function}
\usage{
get_square_error(x)
}
\arguments{
\item{x}{ndp object}
}
\value{
vector of error corresponding to specific iteration's cluster configuration
}
\description{
Calculates Error Distribution under square loss function
}
