% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green_loss.R
\name{green_loss}
\alias{green_loss}
\title{Lau and Green posterior loss function}
\usage{
green_loss(object, truth = NULL, tau = 0.5, a = 1, b = 1)
}
\arguments{
\item{object}{an ndp or bndp object}

\item{truth}{True adjacency matrix}

\item{tau}{loss function tuning parameter}

\item{a}{misclassification penalty parameter}

\item{b}{misclassification penalty parameter - see Green and Lau for details}
}
\value{
list of three objects (1)loss:  the loss corresponding to each clustering 
configuration, (2) best_loss_ix:  The index corresponding to the "best" partition,
and (3) mode: the best partition itself
}
\description{
Lau and Green posterior loss function
}
