% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{green_loss_engine}
\alias{green_loss_engine}
\title{computes green loss function}
\usage{
green_loss_engine(cluster_assignment, pmat, tau)
}
\arguments{
\item{cluster_assignment}{iter_total x J cluster assignment matrix}

\item{pmat}{J x J pairwise probability of co-clustering matrix}
}
\description{
computes green loss function
}
