% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{green_loss_known}
\alias{green_loss_known}
\title{Computes Green and Lau Loss function with known classification}
\usage{
green_loss_known(cluster_assignment, pmat, true_cluster_assignment, a, b)
}
\arguments{
\item{cluster_assignment}{iter_total x J cluster assignment matrix}

\item{pmat}{J x J pairwise probability of co-clustering matrix}

\item{true_cluster_assignment}{J x J true Adjacency Matrix}

\item{a}{mis-classification penalty parameter}

\item{b}{classification penalty parameter}
}
\description{
Computes Green and Lau Loss function with known classification
}
