% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{green_loss_unknown}
\alias{green_loss_unknown}
\title{Computes Green and Lau loss function with unknown classification}
\usage{
green_loss_unknown(cluster_assignment, pmat, tau)
}
\arguments{
\item{cluster_assignment}{iter_total x J cluster assignment matrix}

\item{pmat}{J x J pairwise probability of co-clustering matrix}

\item{tau}{penalty parameter}
}
\description{
Computes Green and Lau loss function with unknown classification
}
