% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benvo-methods.R
\name{groupvo}
\alias{groupvo}
\alias{groupvo.benvo}
\title{Retrieve Grouped Data Structure}
\usage{
groupvo(x, formula)

\method{groupvo}{benvo}(x, formula)
}
\arguments{
\item{x}{benvo object}

\item{formula}{bendr formula}
}
\description{
Retrieve Grouped Data Structure
}
\section{Methods (by class)}{
\itemize{
\item \code{benvo}: return list of grouped data structures
}}

