% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{guessEC50_lm}
\alias{guessEC50_lm}
\title{Guess the logEC50 by fitting a linear model through the datapoints closest to 50% effect concentration}
\usage{
guessEC50_lm(df)
}
\arguments{
\item{df}{dataframe containing effect and log concentration columns}
}
\value{
guess for EC50 value
}
\description{
Guess the logEC50 by fitting a linear model through the datapoints closest to 50% effect concentration
}
