% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{hmc}
\alias{hmc}
\title{Create a HMC model object}
\usage{
hmc(object, warm_up, iter_max, X)
}
\arguments{
\item{object}{model fit object returned from nhpp_gamma Rcpp function}

\item{warm_up}{same as nhpp_hmc}

\item{iter_max}{same as nhpp_hmc}

\item{X}{the design matrix used in the regression}
}
\value{
an hmc model object
}
\description{
Create a HMC model object
}
