% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{inverse.hill}
\alias{inverse.hill}
\title{Inverse hill function}
\usage{
inverse.hill(effect, logEC50, slope)
}
\arguments{
\item{effect}{effect value}

\item{logEC50}{parameter; log10(EC50)}

\item{slope}{parameter; slope}
}
\value{
concentration for given effect
}
\description{
Inverse hill function
}
