% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_measure.R
\name{measures}
\alias{measures}
\title{Base Measures}
\description{
Pros and Cons of different base measures.
}
\details{
One of the "parameters" of the Dirichlet Process is the base measure
which describes the distribution around which the Dirichlet Process is centered.
\code{\link{bend}} allows for two possible base measures: (1) normal and (2) beta.

The normal base measure is faster and offers more efficient samples, but requires 
transforming the distances or radii to the real line via the nonlinear probit link function.
In contrast the beta base measure requires only a linear scaling to keep distances in (0,1)
and may be more appealing for it's lack of "edge effects" in the estimated densities.
}
\seealso{
\link{normal_measure} \link{beta_measure}
}
