% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nd_nhpp.R
\name{nd_nhpp}
\alias{nd_nhpp}
\title{Estimate the nonhomgogenous poisson process intensity function from grouped data using a normal kernel}
\usage{
nd_nhpp(X, r, n_j, mu_0 = 0, kappa_0 = 1, nu_0 = 1, sigma_0 = 1,
  L = 4, K = 4, a_0 = 1, b_0 = 1, a_alpha = 1, b_alpha = 1,
  a_rho = 1, b_rho = 1, iter_max, warm_up, thin = 1,
  include_warmup = FALSE, probit = TRUE, seed = NULL)
}
\arguments{
\item{r}{vector of distances  associated with differing groups}

\item{n_j}{matrix of integers denoting the start and length of each school's associated BEF distances}

\item{mu_0}{mean hyperparameter for mu normal base measure. Default is 0; Normal(0,1).}

\item{kappa_0}{variance hyperparameter for mu normal base measure. Default is 1; Normal(0,1).}

\item{nu_0}{df hyperparameter for sigma inv chisq base measure. Default is 1; InvChisq(1,1);}

\item{sigma_0}{scale hyperparameter for sigma inv chisq base measure. Default is 1; InvChisq(1,1);}

\item{L}{component truncation number}

\item{K}{intensity cluster truncation number}

\item{a_alpha}{hyperparameter for alpha gamma prior}

\item{b_alpha}{hyperparameter for alpha gamma prior}

\item{a_rho}{hyperparameter for rho gamma prior}

\item{b_rho}{hyperparameter for rho gamma prior}

\item{iter_max}{total number of iterations for which to run sampler}

\item{warm_up}{number of iterations for which to burn-in or "warm-up" sampler}

\item{thin}{number of iterations to thin by}

\item{probit}{boolean value to indicate the transformation function.
If TRUE then the probit is used if FALSE then the logit transformation.}

\item{seed}{integer with which to initialize random number generator}
}
\description{
Estimate the nonhomgogenous poisson process intensity function from grouped data using a normal kernel
}
