% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nd_nhpp_fit}
\alias{nd_nhpp_fit}
\title{Estimate the nonhomgogenous poisson process intensity function from grouped data}
\usage{
nd_nhpp_fit(
  r,
  n_j,
  d,
  L,
  K,
  J,
  mu_0,
  kappa_0,
  nu_0,
  sigma_0,
  a_alpha,
  b_alpha,
  a_rho,
  b_rho,
  iter_max,
  warm_up,
  thin,
  seed,
  chain,
  num_posterior_samples
)
}
\arguments{
\item{r}{vector of distances associatd with different BEFs}

\item{n_j}{matrix of integers denoting the start and length of each observations associated BEF distances}

\item{d}{a 1D grid of positive real values over which the differing intensities are evaluated}

\item{L}{component truncation number}

\item{K}{intensity cluster truncation number}

\item{J}{number of rows in r matrix; number of groups}

\item{mu_0}{normal base measure prior mean}

\item{kappa_0}{normal base measure prior variance scale}

\item{nu_0}{inverse chi sqaure base measure prior degrees of freedom}

\item{sigma_0}{inverse chi square base measure prior scale}

\item{a_alpha}{hyperparameter for alpha gamma prior}

\item{b_alpha}{scale hyperparameter for alpha gamma prior}

\item{a_rho}{hyperparameter for rho gamma prior}

\item{b_rho}{scale hyperparameter for rho gamma prior}

\item{iter_max}{total number of iterations for which to run sampler}

\item{warm_up}{number of iterations for which to burn-in or "warm-up" sampler}

\item{thin}{number of iterations to thin by}

\item{seed}{integer with which to initialize random number generator}

\item{chain}{integer chain label}

\item{num_posterior_samples}{the total number of posterior samples after burn in}
}
\description{
Estimate the nonhomgogenous poisson process intensity function from grouped data
}
\references{
Gelman, A., Carlin J., Stern H. and Rubin D. (2004). \emph{Bayesian Data Analysis.} Cambridge University Press,
  Chapman & Hall/CRC.
}
\seealso{
the conjugate normal parameterization in the reference below
}
