% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nd_nhpp_fixed_fit}
\alias{nd_nhpp_fixed_fit}
\title{Estimate the nonhomgogenous poisson process intensity function from grouped data with fixed concentration parameters}
\usage{
nd_nhpp_fixed_fit(r, n_j, d, L, K, J, mu_0, kappa_0, nu_0, sigma_0, alpha,
  rho, iter_max, warm_up, thin, seed, chain, num_posterior_samples)
}
\arguments{
\item{r}{vector of distances associatd with different BEFs}

\item{n_j}{matrix of integers denoting the start and length of each observations associated BEF distances}

\item{d}{a 1D grid of positive real values over which the differing intensities are evaluated}

\item{L}{component truncation number}

\item{K}{intensity cluster truncation number}

\item{J}{number of rows in r matrix; number of groups}

\item{iter_max}{total number of iterations for which to run sampler}

\item{warm_up}{number of iterations for which to burn-in or "warm-up" sampler}

\item{thin}{number of iterations to thin by}

\item{seed}{integer with which to initialize random number generator}

\item{chain}{integer chain label}
}
\description{
Estimate the nonhomgogenous poisson process intensity function from grouped data with fixed concentration parameters
}
