% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndp-methods.R
\name{ndp-methods}
\alias{ndp-methods}
\alias{nsamples.ndp}
\alias{nsamples}
\title{Methods for ndp objects}
\usage{
\method{nsamples}{ndp}(object, ...)
}
\arguments{
\item{object}{ndp object}

\item{...}{Ignored}
}
\description{
Common methods that allow for easy model interrogation
}
