% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndp.R
\name{ndp}
\alias{ndp}
\title{ndp objects}
\usage{
ndp(object)
}
\arguments{
\item{object}{A list provided by the \code{\link{bend}} function}
}
\value{
An ndp object for easy model interrogation
}
\description{
ndp objects
}
