% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndp_fixed.R
\name{ndp_fixed}
\alias{ndp_fixed}
\title{Create a ndp_fixed object}
\usage{
ndp_fixed(object, chains, alpha, rho, J)
}
\arguments{
\item{object}{A list provided by the nd_nhpp function}

\item{chains}{number of chains}
}
\value{
An ndp object
}
\description{
Create a ndp_fixed object
}
