% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nhpp_gamma}
\alias{nhpp_gamma}
\title{Returns draw of beta from posterior distribution via No U-Turn Sampler}
\usage{
nhpp_gamma(warm_up, iter_max, input_X, input_n_j, adapt_delta, seed)
}
\description{
Returns draw of beta from posterior distribution via No U-Turn Sampler
}
