% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp_hmc.R
\name{nhpp_hmc}
\alias{nhpp_hmc}
\title{Estimates the mean number of observations in an Inhomogenous Poisson Process}
\usage{
nhpp_hmc(formula, data, warm_up, iter_max, seed = NULL)
}
\arguments{
\item{formula}{}

\item{seed}{}
}
\description{
Estimates the mean number of observations in an Inhomogenous Poisson Process
}
