% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{nlsfit}
\alias{nlsfit}
\title{two stage nonlinear least squares fit}
\usage{
nlsfit(m.formula, start, fdata)
}
\arguments{
\item{m.formula}{model formula}

\item{start}{vector of start parameters}

\item{fdata}{data frame of data to fit}
}
\value{
nls2 fit object
}
\description{
two stage nonlinear least squares fit
}
