% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_measure.R
\name{normal_measure}
\alias{normal_measure}
\title{Normal Base Measure}
\usage{
normal_measure(mu_0 = 0, nu_0 = 1, sigma_0 = 1, kappa_0 = 1)
}
\arguments{
\item{mu_0}{mean hyperparameter for mu normal base measure. Default is 0; Normal(0,1).}

\item{nu_0}{df hyperparameter for sigma inv chisq base measure. Default is 1; InvChisq(1,1);}

\item{sigma_0}{scale hyperparameter for sigma inv chisq base measure. Default is 1; InvChisq(1,1);}

\item{kappa_0}{variance hyperparameter for mu normal base measure. Default is 1; Normal(0,1).}
}
\description{
Normal Base Measure
}
