% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_plots.R
\name{plot.bndp}
\alias{plot.bndp}
\title{Plot method for bndp objects}
\usage{
\method{plot}{bndp}(x, plotfun = "cluster")
}
\arguments{
\item{x}{an bndp object}

\item{plotfun}{one of "cluster","global", or "network"}
}
\description{
The \code{plot} method for \link{bndp-objects} provides convenient
access to several types of plots useful for visualizing
bndp models using the default settings.
}
