% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ndp}
\alias{plot.ndp}
\title{Plot method for ndp objects}
\usage{
\method{plot}{ndp}(x, plotfun = "cluster", ...)
}
\arguments{
\item{x}{an ndp object}

\item{plotfun}{one of c("cluster","pairs","global","trace" )}

\item{...}{optional arguments for plot function}
}
\description{
The \code{plot} method for ndp-objects provides convenient
access to several types of plots useful for visualizing
ndp models using the default settings.
}
\seealso{
\code{\link{plot_cluster_densities}} \code{\link{plot_pairs}} \code{\link{plot_global_density}} \code{\link{traceplot}}
}
