% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cluster_densities.ndp}
\alias{plot_cluster_densities.ndp}
\title{Plots cluster intensity function densities}
\usage{
\method{plot_cluster_densities}{ndp}(x, p = 0.9, pi_threshold = 0.1)
}
\arguments{
\item{x}{ndp object}

\item{p}{the probability for the credible interval}

\item{pi_threshold}{intensities with probability of assignment greater than pi_threshold are plotted
default is zero}
}
\value{
ggplot plot object
}
\description{
Plots cluster intensity function densities
}
