% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_global_density}
\alias{plot_global_density}
\alias{plot_global_density.ndp}
\title{Plots global density function}
\usage{
plot_global_density(x, p = 0.9, transform = TRUE)

\method{plot_global_density}{ndp}(x, p = 0.9, transform = FALSE)
}
\arguments{
\item{x}{ndp object}

\item{p}{probability mass contained in uncertainty interval}

\item{transform}{boolean denoting whether or not to transform distances}
}
\value{
ggplot plot object
}
\description{
Plots global density function

Plots global density function
}
\section{Methods (by class)}{
\itemize{
\item \code{ndp}: plots global density estimate
}}

