% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_map.ndp}
\alias{plot_map.ndp}
\title{Plots clusters of a given observation with other observations around them within certain
probability of co-clustering}
\usage{
\method{plot_map}{ndp}(x, coords, p = c(0.025, 0.5, 0.9))
}
\arguments{
\item{x}{ndp object}

\item{coords}{dataframe of observation geo coordinates}
}
\value{
ggplot plot object
}
\description{
Plots clusters of a given observation with other observations around them within certain
probability of co-clustering
}
