% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_multi_drc}
\alias{plot_multi_drc}
\title{Plot multiple dose response curves}
\usage{
plot_multi_drc(fit_df, plot.layout = "single")
}
\arguments{
\item{fit_df}{data frame containing the model fits in the column m.fit}

\item{plot.layout}{determines whether all DRC are plotted in one single plot or a plot grid.
one of c('single','multi')}
}
\value{
ggplot object
}
\description{
Plot multiple dose response curves
}
