% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_plots.R
\name{plot_network.bndp}
\alias{plot_network.bndp}
\title{Network Cluster Plot}
\usage{
\method{plot_network}{bndp}(x, mode_label = TRUE)
}
\arguments{
\item{x}{bndp object}

\item{mode_label}{boolean value to indicate that the network graph should be colored with the Mode cluster label assignments}
}
\value{
ggplot plot object
}
\description{
Network Cluster Plot
}
