% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_network}
\alias{plot_network}
\title{Network Cluster Plot}
\usage{
plot_network(x, mode_label = TRUE)
}
\arguments{
\item{x}{ndp object}

\item{mode_label}{boolean value to indicate that the network graph should be colored with the Mode cluster label assignments}
}
\value{
ggplot plot object
}
\description{
Network Cluster Plot
}
