% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_plots.R
\name{plot_pairs.bndp}
\alias{plot_pairs.bndp}
\title{plots pairwise probability clustering plot}
\usage{
\method{plot_pairs}{bndp}(x, sort = FALSE)
}
\arguments{
\item{x}{bndp object}

\item{sort}{boolean asking whether sorting algorithm should be used to sort}
}
\value{
ggplot plot object
}
\description{
plots pairwise probability clustering plot
}
