% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pairs.ndp}
\alias{plot_pairs.ndp}
\title{plots pairwise probability clustering plot}
\usage{
\method{plot_pairs}{ndp}(x, sort = FALSE)
}
\arguments{
\item{x}{ndp object}
}
\value{
ggplot plot object
}
\description{
plots pairwise probability clustering plot
}
