% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pairs}
\alias{plot_pairs}
\title{plots pairwise probability clustering plot}
\usage{
plot_pairs(x, sort = FALSE)
}
\arguments{
\item{x}{ndp object}

\item{sort}{boolean asking whether sorting algorithm should be used to sort
pairwise probablity}
}
\value{
ggplot plot object
}
\description{
plots pairwise probability clustering plot
}
\references{
Rodriguez A., Dunson D, and Gelfand A. \emph{The Nested Dirichlet Process}
Journal of the American Statistical Association 103 (1131-1154).
}
\seealso{
the supplementary section of the reference for the sorting algorithm.
}
