% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_replicate_drc}
\alias{plot_replicate_drc}
\title{Plot a single dose response curve with standard deviation of replicates}
\usage{
plot_replicate_drc(fitObject)
}
\arguments{
\item{fitObject}{bendr fit object}
}
\value{
ggplot object
}
\description{
Plot a single dose response curve with standard deviation of replicates
}
