% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_single_drc}
\alias{plot_single_drc}
\title{Plot a single dose response curve}
\usage{
plot_single_drc(fitObject)
}
\arguments{
\item{fitObject}{drfit result object}
}
\description{
Plot a single dose response curve
}
