% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_traceplots.ndp}
\alias{plot_traceplots.ndp}
\title{Plots traceplots of various groups of parameters}
\usage{
\method{plot_traceplots}{ndp}(x, par = "alpha")
}
\arguments{
\item{x}{ndp object}
}
\description{
Plots traceplots of various groups of parameters
}
