% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{prediction_ci}
\alias{prediction_ci}
\title{Calculate the confidence intervals using the delta method}
\usage{
prediction_ci(
  m.formula,
  x.values,
  curvefit,
  curve.predict,
  dof,
  level = 0.95,
  verbose = FALSE
)
}
\arguments{
\item{m.formula}{model formula}

\item{x.values}{concentration values at which the curve will be plotted}

\item{curvefit}{model fit object from nonlinear least squares}

\item{curve.predict}{predicted effect values at the concentrations specified in x.values}

\item{dof}{degrees of freedom for the inverse student's t distribution ; default is (# measurements)-(number of parameters)}

\item{level}{Level of confidence interval to use (0.95 by default). [0,1]}

\item{verbose}{Print and plot intermediate results for verboseging}
}
\value{
Confidence interval values for all concentrations in x.values
}
\description{
Calculate the confidence intervals using the delta method
}
