% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-and-summary.R
\name{print.ndp}
\alias{print.ndp}
\title{Print method for ndp objects}
\usage{
\method{print}{ndp}(x, digits = 2, ...)
}
\arguments{
\item{x}{a ndp object}

\item{digits}{Number of digits to use for formatting numbers.}

\item{...}{Ignored.}
}
\value{
Returns \code{x}, invisibly.
}
\description{
The \code{print} method for ndp objects displays a compact summary of the
fitted model. See the \strong{Details} section below for descriptions of the
different components of the printed output. For additional summary statistics
and diagnostics use the \code{\link[=summary.ndp]{summary}} method.
}
\details{
\subsection{Point Estimates}{
Point estimates are medians computed from simulations.
}
\subsection{Uncertainty estimates (SD)}{
The standard deviations reported (labeled \code{MAD} in the print output)
are computed from the same set of draws described above and are proportional
to the median absolute deviation (\code{\link[stats]{mad}}) from the median.
Compared to the raw posterior standard deviation, the MAD_SD will be
more robust for long-tailed distributions. These are the same as the values
}
}
\seealso{
\code{\link{summary.ndp}}
}
