% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_print-and-summary.R
\name{print.summary.bndp}
\alias{print.summary.bndp}
\title{Printing NDP Beta Measure Summaries}
\usage{
\method{print}{summary.bndp}(x, digits = max(1, attr(x, "print.digits")),
  ...)
}
\arguments{
\item{x}{An object of class \code{"summary.bndp"}.}

\item{digits}{Number of digits to use for formatting numbers when printing. 
When calling \code{summary}, the value of digits is stored as the 
\code{"print.digits"} attribute of the returned object.}

\item{...}{Currently ignored.}
}
\description{
Printing NDP Beta Measure Summaries
}
