% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-datasets.R
\name{rndpp-datasets}
\alias{rndpp-datasets}
\alias{school_data}
\title{Datasets for rndpp examples}
\format{\describe{
\item{\code{school_data}}{
Simulated data school data
\itemize{
\item \code{school_id}: The school unique identifier
\item \code{distances}: Distances generated by Inhomogenous Poisson Process
\item \code{density}: density cluster indicator
}
}

}}
\description{
Small datasets for use in \pkg{rndpp} examples and vignettes.
}
