% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnhpp.R
\name{rnhpp}
\alias{rnhpp}
\title{Random nonhomogenous poisson process generator}
\usage{
rnhpp(nsim = 1, lambda = function(x) x^2, max = NULL,
  interval = c(0, 1), seed = NULL)
}
\arguments{
\item{nsim}{number of processes to simulate}

\item{lambda}{intensity function (unnormalized)}

\item{max}{optional maximum value of intensity function}

\item{interval}{1D interval overwhich process is realized}

\item{seed}{rng seed initializer}
}
\description{
Random nonhomogenous poisson process generator
}
