% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{square_error}
\alias{square_error}
\title{Computes Square loss with unknown classification}
\usage{
square_error(cluster_assignment, pmat)
}
\arguments{
\item{cluster_assignment}{iter_total x J cluster assignment matrix}

\item{pmat}{J x J pairwise probability of co-clustering matrix}
}
\description{
Computes Square loss with unknown classification
}
