% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_print-and-summary.R
\name{summary.bndp}
\alias{summary.bndp}
\title{Summary method for ndp objects}
\usage{
\method{summary}{ndp}(object, probs, digits = 1, ...)
}
\arguments{
\item{probs}{For models fit using MCMC, 
an optional numeric vector of probabilities passed to 
\code{\link[stats]{quantile}}.}

\item{digits}{Number of digits to use for formatting numbers when printing. 
When calling \code{summary}, the value of digits is stored as the 
\code{"print.digits"} attribute of the returned object.}

\item{...}{Currently ignored.}
}
\value{
The \code{summary} method returns an object of class 
  \code{"summary.ndp"}.
}
\description{
Summaries of parameter estimates and MCMC convergence diagnostics 
(Monte Carlo error, effective sample size, Rhat).
}
