% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_print-and-summary.R
\name{to_latex.bndp}
\alias{to_latex.bndp}
\title{converts summary output to latex table}
\usage{
\method{to_latex}{bndp}(object, digits = 1, caption = "")
}
\description{
converts summary output to latex table
}
