% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-and-summary.R
\name{to_latex.ndp}
\alias{to_latex.ndp}
\title{converts summary output to latex table}
\usage{
\method{to_latex}{ndp}(object, digits = 1, caption = "")
}
\description{
converts summary output to latex table
}
