% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_print-and-summary.R, R/print-and-summary.R
\name{to_latex}
\alias{to_latex}
\title{converts summary output to latex}
\usage{
\method{to_latex}{ndp}(object, digits = 1, caption = "")

\method{to_latex}{ndp}(object, digits = 1, caption = "")
}
\arguments{
\item{object}{an ndp object}

\item{digits}{number of digits to round to}

\item{caption}{self-explanatory}

\item{object}{an ndp object}

\item{digits}{number of digits to round to}

\item{caption}{self-explanatory}
}
\description{
converts summary output to latex

converts summary output to latex
}
