% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{traceplot}
\alias{traceplot}
\alias{traceplot.ndp}
\title{Traceplots of various NDP-NHPP parameters}
\usage{
traceplot(x, par = "alpha")

\method{traceplot}{ndp}(x, par = "alpha")
}
\arguments{
\item{x}{ndp object}

\item{par}{character vector of parameter names, defaults to c("alpha")}
}
\value{
ggplot plot object
}
\description{
Traceplots of various NDP-NHPP parameters
}
\section{Methods (by class)}{
\itemize{
\item \code{ndp}: parameter traceplots
}}

