% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_measure.R
\name{transform_distances}
\alias{transform_distances}
\title{Transform Distances}
\usage{
transform_distances(bm, r)
}
\arguments{
\item{bm}{base measure object}

\item{r}{vector of radii}
}
\description{
Transform Distances
}
\details{
Transforms distances according to which base measure is used.
Either a linear transformation - scaling by the ceiling of the maximum distance 
or a nonlinear transformation via the probit function.
}
