% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_mode.R
\name{assign_mode}
\alias{assign_mode}
\title{Estimates Posterior Mode Cluster Assignment}
\usage{
assign_mode(x, ...)
}
\arguments{
\item{x}{ndp object}

\item{...}{optional arguments}
}
\value{
vector of cluster assignments
}
\description{
Estimates Posterior Mode Cluster Assignment
}
