% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bend.R
\name{bend}
\alias{bend}
\title{Estimate the distribution of (B)uilt (E)nvironment amenities via the (N)ested (D)irichlet Process}
\usage{
bend(
  formula,
  benvo,
  base_measure = normal_measure(),
  L = 4,
  K = 4,
  a_alpha = 1,
  b_alpha = 1,
  a_rho = 1,
  b_rho = 1,
  iter_max = 5000,
  burn_in = floor(iter_max/2),
  thin = 1,
  fix_concentration = FALSE,
  seed = NULL
)
}
\arguments{
\item{formula}{bendr formula see the details}

\item{benvo}{\href{https://apeterson91.github.io/rbenvo}{\code{Benvo}} object}

\item{base_measure}{"bm" object from one of the base measure functions. See \link{measures}.}

\item{L}{component truncation number}

\item{K}{intensity cluster truncation number}

\item{a_alpha}{shape hyperparameter for alpha gamma prior. Fixed alpha if fix_concentration = TRUE.}

\item{b_alpha}{scale hyperparameter for alpha gamma prior}

\item{a_rho}{shape hyperparameter for rho gamma prior. Fixed rho if fix_concentration = TRUE.}

\item{b_rho}{scale hyperparameter for rho gamma prior}

\item{iter_max}{total number of iterations for which to run sampler}

\item{burn_in}{number of iterations to discard as sampler approaches the stationary distribution}

\item{thin}{number of iterations to thin by}

\item{fix_concentration}{boolean value indicating whether or not concentration parameters should be kept fixed. Currently only implimented for alpha base measure}

\item{seed}{integer with which to initialize random number generator}
}
\description{
Estimate the distribution of (B)uilt (E)nvironment amenities via the (N)ested (D)irichlet Process
}
\details{
Fits the Nested Dirichlet Process to data as specified in the formula and hyperparameter settings. Note that a 'bendr formula' simply has the grouping id on the left side of the tilde and the built environment feature name on the right hand side. 
No intercepts, or other terms are accepted or needed.
}
\seealso{
The introductory \href{https://apeterson91.github.io/bendr/articles/Introduction.html}{vignette} for an example on how this function is used.
}
