% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calc_ecx}
\alias{calc_ecx}
\title{calculate ECx}
\usage{
calc_ecx(x, slope, ec50)
}
\arguments{
\item{slope}{slope of the hill curve}

\item{ec50}{Effective concentration at 50%}
}
\value{
ECx value
}
\description{
calculate ECx
}
