% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNTC.R
\name{findNTC}
\alias{findNTC}
\title{Find the non-toxic concentration.}
\usage{
findNTC(bendrObj)
}
\arguments{
\item{bendrObj}{A bendr result object}
}
\value{
The absolute (non-logged) non-toxic concentration value
}
\description{
If you use the result of this calculation in your research, please cite Stadnicka-Michalak et al.
}
\examples{
require(tibble)
drc.formula = effect ~ 100 / (1 + 10^((logEC50-logconc) * slope))
rdata = as_tibble(
data.frame(
   conc=c(0.1,1,10,100,1000,5000),
   replicateA=c(1,12,27,65,88,100),
   replicateB=c(2,16,21,62,81,100),
   replicateC=c(4,15,21,61,85,96)
 ))

fitObjectRep = fitdr_replicates(drc.formula,rdata,2:4,conc)
ntc = findNTC(fitObjectRep)

}
