% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fitdr_replicates}
\alias{fitdr_replicates}
\title{Dose response curve fit with replicates}
\usage{
fitdr_replicates(m.formula, fdata, effectColumns, concColumn,
  level = 0.95, start = vector(), ci_method = "delta", verbose = F)
}
\arguments{
\item{m.formula}{model formula}

\item{fdata}{dataframe with data to fit}

\item{effectColumns}{effect column names or indices in data frame}

\item{concColumn}{concentration column name in dataframe}

\item{level}{confidence level}

\item{start}{starting values}

\item{ci_method}{method for confidence interval calculation, one of ["delta","bootstrap"]}

\item{verbose}{print fitting intermediate results and confidence interval calculation intermediate  results}
}
\description{
Dose response curve fit with replicates
}
