% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fitdr_replicates}
\alias{fitdr_replicates}
\title{Dose response curve fit with replicates}
\usage{
fitdr_replicates(m.formula, fdata, effectColumns, concColumn,
  level = 0.95, start = vector(), verbose = F, debug = F)
}
\arguments{
\item{verbose}{}
}
\description{
Dose response curve fit with replicates
}
