% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green_loss.R
\name{green_loss.ndp}
\alias{green_loss.ndp}
\title{Lau and Green posterior loss function}
\usage{
\method{green_loss}{ndp}(object, tau = 0.5)
}
\arguments{
\item{tau}{loss function tuning parameter}
}
\description{
Lau and Green posterior loss function
}
