% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{groupvo}
\alias{groupvo}
\alias{groupvo,Benvo-method}
\title{Retrieve Grouped Data Structure}
\usage{
groupvo(x, formula)

\S4method{groupvo}{Benvo}(x, formula)
}
\arguments{
\item{x}{benvo object}

\item{formula}{bendr formula}
}
\description{
Retrieve Grouped Data Structure

Grouped
}
\section{Methods (by class)}{
\itemize{
\item \code{Benvo}: return list of grouped data structures
}}

