% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nd_nhpp_fixed.R
\name{nd_nhpp_fixed}
\alias{nd_nhpp_fixed}
\title{Estimate the nonhomgogenous poisson process intensity function from grouped data using a normal kernel}
\usage{
nd_nhpp_fixed(X, r, n_j, mu_0 = 0, kappa_0 = 1, nu_0 = 1,
  sigma_0 = 1, alpha = 1, rho = 1, L = 4, K = 4, iter_max,
  warm_up, thin = 1, seed = NULL)
}
\arguments{
\item{r}{vector of distances  associated with differing groups}

\item{n_j}{matrix of integers denoting the start and length of each school's associated BEF distances}

\item{mu_0}{mean hyperparameter for mu normal base measure. Default is 0; Normal(0,1).}

\item{kappa_0}{variance hyperparameter for mu normal base measure. Default is 1; Normal(0,1).}

\item{nu_0}{df hyperparameter for sigma inv chisq base measure. Default is 1; InvChisq(1,1);}

\item{sigma_0}{scale hyperparameter for sigma inv chisq base measure. Default is 1; InvChisq(1,1);}

\item{alpha}{outer DP concentration parameter}

\item{rho}{inner DP concentration parameter}

\item{L}{component truncation number}

\item{K}{intensity cluster truncation number}

\item{iter_max}{total number of iterations for which to run sampler}

\item{warm_up}{number of iterations for which to burn-in or "warm-up" sampler}

\item{thin}{number of iterations to thin by}

\item{seed}{integer with which to initialize random number generator}
}
\description{
Estimate the nonhomgogenous poisson process intensity function from grouped data using a normal kernel
}
