% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndp.R
\name{ndp}
\alias{ndp}
\title{Create a ndp object}
\usage{
ndp(object, chains)
}
\arguments{
\item{object}{A list provided by the nd_nhpp function}

\item{chains}{number of chains}
}
\value{
An ndp object
}
\description{
Create a ndp object
}
