% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp_hmc.R
\name{nhpp_hmc}
\alias{nhpp_hmc}
\title{Estimates the mean number of observations in an Inhomogenous Poisson Process}
\usage{
nhpp_hmc(formula, data, warm_up, iter_max, seed = NULL)
}
\arguments{
\item{data}{data.frame from which to extract outcome and covariates}

\item{warm_up}{number of iterations in which to tune HMC step-size, these will be discarded}

\item{iter_max}{total number of samples for which to run sampler}

\item{seed}{}
}
\description{
currently applies a N(0,sigma=3) prior to all regression coefficients
}
