% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ndp}
\alias{plot.ndp}
\title{Plot method for ndp objects}
\usage{
\method{plot}{ndp}(x, plotfun = "cluster", ...)
}
\arguments{
\item{x}{an ndp object}

\item{plotfun}{one of "cluster","global", or "network"}

\item{...}{ignored}
}
\description{
The \code{plot} method for ndp-objects provides convenient
access to several types of plots useful for visualizing
ndp models using the default settings.
}
