% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ndp}
\alias{plot.ndp}
\title{Plot method for ndp objects}
\usage{
\method{plot}{ndp}(x, plotfun = "intervals", pars = NULL,
  regex_pars = NULL, ...)
}
\arguments{
\item{x}{an ndp object}
}
\description{
The \code{plot} method for \link{ndp-objects} provides convenient 
access to several types of plots useful for visualizing
ndp models
}
