% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cluster_densities.ndp}
\alias{plot_cluster_densities.ndp}
\title{Plots cluster intensity function densities}
\usage{
\method{plot_cluster_densities}{ndp}(
  x,
  p = 0.9,
  pi_threshold = 0.1,
  switch = "facet",
  transform = FALSE
)
}
\arguments{
\item{x}{ndp object}

\item{p}{the probability for the credible interval}

\item{pi_threshold}{intensities with probability of assignment greater than pi_threshold are plotted
default is zero}

\item{switch}{one of two character values "facet" or "color" denoting whether to separate clusters by facet or by color}

\item{transform}{boolean value denoting whether or not to transform the distances: FALSE indicates values
are shown on the [0,R) domain, while TRUE indicates they are shown on the real line.}
}
\value{
ggplot plot object
}
\description{
Plots cluster intensity function densities
}
