% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_plots.R
\name{plot_global_density.bndp}
\alias{plot_global_density.bndp}
\title{Plots global density function - for monitoring convergence}
\usage{
\method{plot_global_density}{bndp}(x, p = 0.9, r = NULL)
}
\arguments{
\item{x}{bndp object}

\item{p}{probability mass contained in uncertainty interval}
}
\value{
ggplot plot object
}
\description{
Plots global density function - for monitoring convergence
}
