% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_network.ndp}
\alias{plot_network.ndp}
\title{Network Cluster Plot}
\usage{
\method{plot_network}{ndp}(x, sample = TRUE)
}
\arguments{
\item{x}{ndp object}

\item{sample}{boolean indicating random subsample will be taken for plot}
}
\value{
ggplot plot object
}
\description{
Network Cluster Plot
}
