% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pairs}
\alias{plot_pairs}
\alias{plot_pairs.default}
\title{plots pairwise probability clustering plot}
\usage{
plot_pairs(x, sample = NULL, sort = FALSE)

\method{plot_pairs}{default}(x, sample = NULL, sort = FALSE)
}
\arguments{
\item{x}{object with upper triangle pairwise probability matrix list entry}

\item{sample}{positive integer denoting size of random sample for which the pairwise graph can be subset - useful for speeding up graphing}

\item{sort}{boolean asking whether sorting algorithm should be used to sort
pairwise probablity}
}
\value{
ggplot plot object
}
\description{
plots pairwise probability clustering plot

plots pairwise probability clustering plot
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: constructs pairwise probability heat map
}}

\references{
Rodriguez A., Dunson D, and Gelfand A. \emph{The Nested Dirichlet Process}
Journal of the American Statistical Association 103 (1131-1154).
}
\seealso{
the supplementary section of the reference for the sorting algorithm.
}
