% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_print-and-summary.R
\name{summary.bndp}
\alias{summary.bndp}
\title{Summary method for bndp objects}
\usage{
\method{summary}{bndp}(object, digits = 1, ...)
}
\arguments{
\item{object}{a bndp object}

\item{digits}{Number of digits to use for formatting numbers when printing. 
When calling \code{summary}, the value of digits is stored as the 
\code{"print.digits"} attribute of the returned object.}

\item{...}{Currently ignored.}
}
\value{
The \code{summary} method returns an object of class 
  \code{"summary.bndp"}.
}
\description{
Summaries of parameter estimates and MCMC convergence diagnostics 
(Monte Carlo error, effective sample size).
}
